requirements.txt
# Core dependencies
httpx[http2]>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
typer[all]>=0.9.0
rich>=13.7.0

# Async support
asyncio>=3.4.3
aiofiles>=23.2.1

# Database
aiosqlite>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
pre-commit>=3.6.0

# API documentation
pydantic-openapi-schema>=1.3.0

# CLI v2 dependencies
pyyaml>=6.0.1
watchdog>=3.0.0
schedule>=1.2.0
click>=8.1.7  # For typer

# Error handling and monitoring
psutil>=5.9.0  # For system resource monitoring
tabulate>=0.9.0  # For error analysis CLI

# Cache system dependencies
lz4>=4.3.2  # For LZ4 compression support